+++ packages/components-data/src/index.ts
@@ -26,3 +26,4 @@
 
 export * from './DataProvider'
 export * from './hooks'
+export * from './utils'