+++ packages/components/src/Calendar/index.ts
@@ -26,3 +26,4 @@
 
 export * from './Calendar'
 export * from './types'
+export * from './utils'