+++ packages/components/src/Calendar/utils/index.ts
@@ -25,5 +25,7 @@
  */
 
 export * from './formatDateString'
+export * from './getLocale'
 export * from './parseDateFromString'
 export * from './rangeTypeStyle'
+export * from './TimeFormat'