+++ packages/filter-components/src/Filter/components/AdvancedFilter/components/DateFilter/components/RelativeTimeframes/components/RelativeTimeframeCustom/RelativeTimeframeCustom.tsx
@@ -24,9 +24,8 @@
 
  */
 
-import { InputDateRange } from '@looker/components-date'
+import { Box2, InputDateRange } from '@looker/components'
 import React from 'react'
-import styled from 'styled-components'
 import type { DayRange } from '../../../../types/day_range'
 import type { RelativeTimeframeModel } from '../../../../types/relative_timeframe_types'
 import { FILTERS_DATE_FORMAT } from '../../../../utils/format_date'
@@ -36,11 +35,16 @@ interface RelativeTimeframeCustomProps {
   onCustomChange: (range: DayRange) => void
 }
 
+const defaultValue = {
+  from: new Date(Date.now()),
+  to: new Date(Date.now()),
+}
+
 export const RelativeTimeframeCustom = ({
   value,
   onCustomChange,
 }: RelativeTimeframeCustomProps) => {
-  const range = typeof value === 'string' ? undefined : value
+  const range = typeof value === 'string' ? defaultValue : value
 
   const handleCustomChange = (d: Partial<DayRange> = {}) => {
     const newDateRange = {
@@ -52,7 +56,7 @@ export const RelativeTimeframeCustom = ({
   }
 
   return (
-    <InputWrapper>
+    <Box2 p="u3">
       <InputDateRange
         // hardcoding 'yyyy/MM/dd' for legacy reasons
         // eventually this should be replaced with the user's preferred locale
@@ -60,14 +64,6 @@ export const RelativeTimeframeCustom = ({
         value={range}
         onChange={handleCustomChange}
       />
-    </InputWrapper>
+    </Box2>
   )
 }
-
-const InputWrapper = styled.div`
-  padding: ${({ theme }) => theme.space.small};
-  & > div {
-    display: flex;
-    flex-direction: column;
-  }
-`